backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
dramatiq[redis]==1.17.1
dramatiq-abort[redis]==1.2.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0 